changed to use a class based background for the main function
